+++ pkg/services/object/server.go
@@ -354,7 +354,6 @@ func (x *putStream) forwardRequest(req *protoobject.PutRequest) error {
 		var p putsvc.PutInitPrm
 		p.WithCommonPrm(cp)
 		p.WithObject(obj)
-		p.WithCopiesNumber(v.Init.CopiesNumber)
 		p.WithRelay(x.sendToRemoteNode)
 		if err = x.base.Init(&p); err != nil {
 			return fmt.Errorf("could not init object put stream: %w", err)